edit struct naming
