+++ crates/cubek-pool/src/definition/base.rs
@@ -1,3 +1,4 @@
+use crate::definition::{AdaptiveAvgPoolOptions, AvgPoolOptions, MaxPoolOptions};
 use cubecl::zspace::Shape;
 
 #[derive(Clone, Debug)]
@@ -42,75 +43,4 @@ impl<const N: usize> From<AdaptiveAvgPoolOptions<N>> for PoolMode<N> {
     fn from(options: AdaptiveAvgPoolOptions<N>) -> Self {
         PoolMode::AdaptiveAvg(options)
     }
-}
-
-#[derive(Clone, Debug)]
-pub struct MaxPoolOptions<const N: usize> {
-    pub window: SlidingWindow<N>,
-    pub dilation: [usize; N],
-}
-
-impl<const N: usize> MaxPoolOptions<N> {
-    pub fn new(
-        kernel_size: [usize; N],
-        stride: [usize; N],
-        padding: [usize; N],
-        dilation: [usize; N],
-        ceil_mode: bool,
-    ) -> Self {
-        Self {
-            window: SlidingWindow {
-                kernel_size,
-                stride,
-                padding,
-                ceil_mode,
-            },
-            dilation,
-        }
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct AvgPoolOptions<const N: usize> {
-    pub window: SlidingWindow<N>,
-    pub count_include_pad: bool,
-}
-
-impl<const N: usize> AvgPoolOptions<N> {
-    pub fn new(
-        kernel_size: [usize; N],
-        stride: [usize; N],
-        padding: [usize; N],
-        ceil_mode: bool,
-        count_include_pad: bool,
-    ) -> Self {
-        Self {
-            window: SlidingWindow {
-                kernel_size,
-                stride,
-                padding,
-                ceil_mode,
-            },
-            count_include_pad,
-        }
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct AdaptiveAvgPoolOptions<const N: usize> {
-    pub output_size: [usize; N],
-}
-
-impl<const N: usize> AdaptiveAvgPoolOptions<N> {
-    pub fn new(output_size: [usize; N]) -> Self {
-        Self { output_size }
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct SlidingWindow<const N: usize> {
-    pub kernel_size: [usize; N],
-    pub stride: [usize; N],
-    pub padding: [usize; N],
-    pub ceil_mode: bool,
 }
