+++ crates/cubek-pool/src/definition/mod.rs
@@ -1,5 +1,7 @@
 mod base;
 mod error;
+mod options;
 
 pub use base::*;
 pub use error::*;
+pub use options::*;
